simplify sample
