+++ app/build.gradle
@@ -51,6 +51,7 @@ dependencies {
   implementation 'androidx.core:core-ktx:1.9.0'
   implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.6.1'
   implementation 'androidx.activity:activity-compose:1.7.0'
+  implementation "androidx.navigation:navigation-compose:2.5.3"
   implementation "androidx.compose.ui:ui:$compose_version"
   implementation "androidx.compose.ui:ui-tooling-preview:$compose_version"
   implementation "androidx.compose.foundation:foundation:$compose_version"
